Sir Bilal Khan Assignments/Assignment 03 Build a Smart Customer Support Bot using OpenAI Agent SDK/my_agent/human_agent.py
# agents/human_agent.py
from openai import Agent

HumanAgent = Agent(
    name="Human Support Agent",
    instructions="You are a human agent. Take over when escalation is required.",
    model="gemini-1.5-pro"
)
